End Project Signed-off-by: GianmarcoSetzu1 <setzu.gianmarco@yahoo.it>

+++ Assets/UI in Game/UI Scripts/BoostBar.cs
@@ -38,7 +38,7 @@ public class BoostBar : MonoBehaviour
     private void OnGUI()
     {
         
-        GUI.Box(new Rect(Screen.width - 70, Screen.height - 40, 60, 20), boost + " / " + boostMax);
+        GUI.Box(new Rect(Screen.width - 70, Screen.height - 40, 60, 20), (int)boost + " / " + boostMax);
         GUI.Box(new Rect(Screen.width-50, Screen.height - 50, 20, -Screen.height / 3), "");
 
         if(boost>0)
